Move cacheMeta storage from module-scope WeakMaps to Atom properties
Store __cacheMeta and __cacheMetaSelector directly on the Atom object
instead of in separate module-scope WeakMaps. Eliminates cacheMetaAtoms.ts
and the hidden global registries. CacheMeta type moves to types/Atom.ts.
