+++ packages/valdres/src/types/Atom.ts
@@ -3,6 +3,15 @@ import type { AtomOnInit } from "./AtomOnInit"
 import type { AtomOnSet } from "./AtomOnSet"
 import type { EqualFunc } from "./EqualFunc"
 import type { Reactive } from "./Reactive"
+import type { Selector } from "./Selector"
+
+export type CacheMeta = {
+    isRevalidating: boolean
+    lastSuccessAt: number
+    maxAge: number
+    staleWhileRevalidate?: number
+    staleIfError?: number
+}
 
 export type Atom<Value = unknown> = {
     equal: EqualFunc<Value>
@@ -15,4 +24,6 @@ export type Atom<Value = unknown> = {
     mutable?: boolean
     staleWhileRevalidate?: Reactive<number>
     staleIfError?: Reactive<number>
+    __cacheMeta?: Atom<CacheMeta | null>
+    __cacheMetaSelector?: Selector<CacheMeta | null>
 }
